+++ packages/fast-components-react-msft/app/app.tsx
@@ -15,6 +15,7 @@ import {
     DesignSystem,
     DesignSystemDefaults,
 } from "@microsoft/fast-components-styles-msft";
+import designSystemSchema from "../../fast-components-styles-msft/src/design-system/design-system.schema.json";
 import { Plugin, PluginProps } from "@microsoft/fast-tooling-react";
 import { HypertextClassNameContract } from "@microsoft/fast-components-class-name-contracts-base";
 import { glyphBuildingblocks } from "@microsoft/fast-glyphs-msft";
@@ -139,6 +140,10 @@ export default class App extends React.Component<{}, AppState> {
                 activeTheme={this.getThemeById(this.state.theme)}
                 showTransparencyToggle={true}
                 styleEditing={true}
+                designSystemEditing={{
+                    data: DesignSystemDefaults,
+                    schema: designSystemSchema,
+                }}
             >
                 <SiteTitle slot={"title"}>
                     <SiteTitleBrand>FAST</SiteTitleBrand> Documentation